Use slightly bigger partials
The current workload does very little string concatenating, but does
a ton of partial lookup. This heavily favors implementations with static
lookup.

There's always the debate of what is representative and what isn't,
but I think it'd be worth benchmarking actual templating a bit more,
hence why I propose to not reduce the number of lookups, but to
increase the amount of concatenation.

This bring the rendered template to about 15kB. For comparison
the Hacker News homepage, a very lightweight website by modern standards,
is just shy of 8KB.

For another point of comparison, https://github.com/joeldrapper/view-layer-benchmarks
is 62KB.